Implement exact BigInt conversions and frozen empty object identity

+++ src/Tsonic.CSharp.Js/JSArrayStatics.cs
@@ -80,7 +80,7 @@ namespace Tsonic.CSharp.Js
             return JSArray<T>.of(items);
         }
 
-        public static JSArray<T> withLength<T>(int length)
+        public static JSArray<T> withLength<T>(double length)
         {
             return new JSArray<T>(length);
         }
